refactor(vscode-extension): streamline agent prompt handling with utility function
- Replaced direct diagnostic injection logic in callCursorAgent and callWindsurfAgent with a new utility function, injectPromptDiagnosticWithCallback.
- Updated prompt messages for clarity and consistency.
- Enhanced error handling and resource management in the new utility function.
